Move routes out of the sever.js file
